video 41

+++ src/main/java/vn/hoidanit/jobhunter/controller/UserController.java
@@ -1,7 +1,11 @@
 package vn.hoidanit.jobhunter.controller;
 
 import java.util.List;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -11,6 +15,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import vn.hoidanit.jobhunter.domain.User;
 import vn.hoidanit.jobhunter.service.UserService;
+import vn.hoidanit.jobhunter.service.error.IdInvalidException;
 
 @RestController
 public class UserController {
@@ -21,35 +26,41 @@ public class UserController {
         this.userService = userService;
     }
 
-    @PostMapping("/user")
-    public User createNewUserPost(@RequestBody User user) {
+    @PostMapping("/users")
+    public ResponseEntity<User> createNewUserPost(@RequestBody User user) {
         User newUser = userService.createNewUser(user);
-
-        return newUser;
+        return ResponseEntity.status(HttpStatus.CREATED).body(newUser);
     }
 
-    @DeleteMapping("/user/{id}")
-    public String deleteUser(@PathVariable Long id) {
+    @DeleteMapping("/users/{id}")
+    public ResponseEntity<String> deleteUser(@PathVariable Long id) throws IdInvalidException {
+
+        if (id > 1500) {
+            throw new IdInvalidException(" ID khong lon hon 1500");
+        }
         userService.deleteUser(id);
 
-        return "delete successfully";
+        return ResponseEntity.ok("delete successfully");
+        // return ResponseEntity.status(HttpStatus.OK).body("delete successfully");
     }
 
-    @GetMapping("/user/{id}")
-    public User getUser(@PathVariable Long id) {
-
-        return userService.findById(id);
+    @GetMapping("/users/{id}")
+    public ResponseEntity<User> getUser(@PathVariable Long id) {
+        User newUser = userService.findById(id);
+        return ResponseEntity.status(HttpStatus.OK).body(newUser);
     }
 
-    @GetMapping("/user")
-    public List<User> getAllUser() {
-        return userService.findAllUser();
+    @GetMapping("/users")
+    public ResponseEntity<List<User>> getAllUser() {
+        List<User> listUser = userService.findAllUser();
+        return ResponseEntity.status(HttpStatus.OK).body(listUser);
     }
 
-    @PutMapping("/user")
-    public User updateUser(@RequestBody User user) {
-        User newUser = userService.createNewUser(user);
+    @PutMapping("/users")
+    public ResponseEntity<User> updateUser(@RequestBody User user) {
+        User newUser = userService.updateUser(user);
 
-        return newUser;
+        return ResponseEntity.status(HttpStatus.OK).body(newUser);
     }
+
 }
